Expose function to get defaults dict
shade needs sane defaults for some of these too, but it's entirely
legitimate to use shade without having first created an os_client_config
object. Adding a function to get a copy of the defaults dict allows it to be
consumed as a single source or truth for default values.

Change-Id: I616d9492c7e0f53c48519cc8dacf3dfbd0082e36